feat(food-details): separate intent class in separate file .
