+++ lib/modules/food_details/ui/view_model/food_details_intent.dart
@@ -1,0 +1,13 @@
+sealed class FoodDetailsIntent {}
+
+class UpdateVideoStateIntent extends FoodDetailsIntent {
+  bool isPlaying;
+
+  UpdateVideoStateIntent(this.isPlaying);
+}
+
+class GetFoodDetailsIntent extends FoodDetailsIntent {
+  String id;
+
+  GetFoodDetailsIntent(this.id);
+}
