+++ lib/modules/food_details/ui/widgets/play_button.dart
@@ -3,6 +3,7 @@ import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
 
 import '../view_model/food_details_cubit.dart';
+import '../view_model/food_details_intent.dart';
 
 class PlayButton extends StatelessWidget {
   const PlayButton({super.key});
